fcft: from_font_set: don't iterate the font set
Fail immediately if the indicated font cannot be loaded. It is up to
the caller to iterate the font list.

This fixes a regression introduced in
a17a9788b55a5a47615522e9d7c338e2430cb401, where we ignore fonts that
does not have the indicated glyph.

Doing so meant we now started iterating the font set list where we
previously had not. As a result, trying a user specified fallback font
could result in one **its** fallback fonts being loaded instead, which
is **not** what we wanted.

Previously, we would have loaded the user specified fallback font, and
then failed when trying to actually load the glyph.