fcft: re-work subpixel antialias handling
Again, the problem is that one cannot do subpixel antialiasing on top
of transparent surfaces.

The original solution had the user call
font_enable_subpixel_antialias() after instantiating the font. Then,
when rendering a glyph, we checked if it was enabled or not, and if
not, tried to reset the freetype render flags.

The problem with this is that we don't know what to reset
it *to*. FT_RENDER_MODE_NORMAL? Or FT_RENDER_MODE_LIGHT? Or perhaps
FT_RENDER_MODE_MONO?

Solve this by calculating render flags for *both* the subpixel and
non-subpixel cases, and let the user choose whether to enable subpixel
antialiasing in font_glyph_for_wc().