fcft: pattern_from_font_with_adjusted_size: require a ":" before {pixel,}size
When replacing the "size" and "pixelsize" attributes in the pattern
string, require them to be prefixed by a ':'.

This fixed an issue where something like this:

  font_name;size=12:dpi=200

gets translated to:

  font_name;dpi=200:size=12.5

In the first case, 'size' is actually treated as a part of the font
name. In the transformed pattern, 'dpi' is instead treated as part of
the font name.

Since 'dpi' often is added by the application, rather than the user,
this leads to unexpected behavior if the user misspelled the font
string.