fcft: scale glyph's width/height
With a pixel fixup size != 1.0, we apply a scale transform to the
glyph. However, the width/height we presented to the caller wasn't
scaled.

For large -> small fixup factors, this didn't matter, but for small ->
large factors it caused the caller to render only a small part of the
scaled glyph.