fcft: use FcCharSetHasChar() to determine whether font has glyph or not
Don't prevent loading a font just because it doesn't have the
requested glyph.

While it saves memory, it costs a lot of performance when loading lots
of glyphs that doesn't exist in the primary font, or in the first
fallback font(s), as it causes it to load and discard fallback fonts
over and over and over again.

Instead, load the font, and keep a reference to the charset. Then,
when loading a glyph from a font, use FcCharSetHasChar() instead of
FT_Get_Char_Index().