fcft: handle multiple threads instantiating the same font at the same time
We don't want multiple font cache entries with the same hash (i.e. the
same font instantiated and cached multiple times).

Previously, this could happen since the a new cache entry wasn't
pushed until after a font had been instantiated, and the font cache
lock was released in between.

Now, if a font cannot be found in the cache, we reserve an entry while
still holding the lock.

Then, after instantiating the font, we just update this reserve entry.

Other threads that sees this reserved entry will, instead of
instantiating the same font themselves, wait for the first thread to
finish.

This is done by adding a cond variable to the cache entry struct. The
original thread, the one that actually instantiates the font, signals
this cond variable after it has updated the reserved cache entry.

fcft_size_adjust() has been updated in a similar manner.

fcft_destroy() has been updated to deref the font, and possibly remove
it from the cache, while holding the font cache lock.