fcft_size_adjust: use a thread pool to instantiate fallback fonts
fcft_size_adjust() does what it does by getting the FC_SIZE attribute
from the FontConfig compiled pattern, adjust it by the specified
amount and then re-write the string pattern.

This in itself isn't super ugly. The problem is the user configured
fallback fonts. If they haven't been loaded we don't
*have* a compiled pattern. Thus we need to instantiate them. This is
slow. For this reason, that is done by a pool of worker threads.

Note that if a fallback font *has* already been loaded, then we have
the pattern and don't need to instantiate the fallback font
again. This we do here, in the main thread.