fcft: font cache: simplify locking semantics
This removes the per-cache entry lock, previously used in combination
with the per-cache entry cond variable used to signal font
instantiation completion to a waiting thread.

Instead, we now use the regular cache lock.

The reason this wasn't done before is that this requires us to release
the font cache lock. This means there's a race in this, granted,
contrived example:

Two threads want to instantiate the same font. Thread A gets there
first, and B sees this and waits for A to instantiate the font.

A does this, and signals this to B. However, before B gets to bump the
fonts ref counter, A now calls fcft_destroy(). Since B hasn't yet
bumped the ref counter, the font is free:d and removed from the
cache. B now has an invalid cache reference and will crash.

The solution is to make B _tell_ A that it wants a reference, _before_
releasing the cache lock.

Thus, when A updates the reserved cache entry with the newly
instantiated font, _it_ bumps the ref counter _for_ B, and for any
other thread(s) that also wants a reference to it.

Now, when A proceeds with fcft_destroy(), the ref counter doesn't
reach zero and font isn't free:d.

All B has to do now is wait for the signal from A, and then it's done
- no need to bump the ref counter.