fcft_glyph_rasterize: disable usage of LCD filter
This is a feature that is disabled by default in
Freetype. Furthermore, it has been superseded by Freetypes own,
non-patent troubled, subpixel rendering.

Still, it doesn't hurt to have the call there, does it?

The problem is that the filter is per library instance. Not per
font. This means we need to set it *each* time we want to render a
glyph.

Since we can render glyphs, from different fonts, in parallel, we need
to hold a lock while setting the LCD filter *and* while rendering the
glyph.

This is bad for performance.