fcft_precompose: new function, pre-composes two wide characters into one
For example, 'a' and '\u0301' can be combined to 'á'.

If the two wide characters cannot be combined, (wchar_t)-1 is
returned.

This function will also tell the caller whether the wide characters
are from the primary font or not.

A caller can for example determine to render the base + combining
character individually if they both are from the primary font but not
the pre-composed character.