fcft: improved hash function - reduces collisions
I'm well aware this still is a pretty bad hash function, but it is
still vastly better than what we had before.

It is important that the hash function is fast; hence I hope this is
"good enough", as I don't want to make it any more complicated.

This fixes a performance problem when loading *lots* of consecutive
chars (for example, when loading large sets of Unicode blocks).

In particular, this fixes a performance problem in foot with the
`notcurses`'s _allglyphs_ demo.