fcft: refactor: change how we represent the primary and fallback fonts
Previously, fallback fonts where represented using a top-level
font_priv struct. I.e. they were the same thing as a primary font,
except some fields of the struct was never initialized (such as the
glyph cache).

Now, we differentiate between the top-level struct, and actual font
instance data.

The top-level struct is now much simpler; it contains the user visible
"public" portion (which is basically the metrics of the primary font),
the glyph cache and a list of fallback fonts. Most notably, the
primary font is the first fallback font.

When we instantiate a new fcft font, we instantiate the FontConfig
patterns for *all* fonts: the primary, the user-configured fallbacks,
and the fontconfig provided fallback list.

We also extract the charset for each font. These are stored in a
fallback struct.

The fallback struct also contains a font instance pointer. This
pointer is initially NULL, for all fallbacks except the primary font.

When looking up a glyph, we simply loop the fallback list (remember,
the primary font is the first one in the list). If the charset
indicates the font contains the requested glyph, we check if the font
has been loaded or not. If not, we instantiate it. This is where a
FreeType 'face' is instantiated, and all the load and render flags are
calculated. I.e. the "heavy" parts of font loading.

If we *fail* to instantiate a font, it is *removed* from the fallback
list. This is to ensure we never try to instantiate it again, since
that will be a performance bottle neck.

Once the font has been loaded, we load the glyph from it. And here we
trust the fontconfig provided charset information; if the font for
some reason does *not* contain the glyph, we'll end up loading the
"fallback" ("unknown") glyph from that font - we will *not* continue
with the next fallback.