fcft: avoid asan warning when bitmap->buffer is NULL
This can happen with e.g. the glyph for the space character.

In this case, everything about the glyph *is* valid. It's just that it
doesn't have a pixmap.

We don't want to treat it as an invalid glyph since the application
may be interested in its metrics, for example.

Also, applications typically assume that a valid glyph always has a
non-NULL pixmap.

So keep code mostly as-is, but avoid calling memcpy() with a
NULL-pointer. This silences asan warnings.