fcft: glyph_rasterize: fix race that sometimes triggered a crash
When rasterizing a glyph we first look it up in the glyph cache. If
this succeeds, we return the glyph from the cache entry.

The code is written such that a cache lookup returns a *pointer* to
the entry. This is to allow the cache to be updated when a glyph
was *not* found and needs to be rendered.

However, we released the font's lock *before* de-referencing the entry
pointer. This still worked in most cases.

But, if another thread triggered a cache resize operation, the entry
pointer in the first thread would no longer be valid.

The fix is to de-reference the cache entry (i.e. to get a direct
pointer to the glyph) *before* releasing the lock.