fcft: rasterize: protect glyph cache with a rwlock
Using an rwlock instead of a mutex to protect the glyph cache means
multiple threads can do a read-only lookup in the cache.

When a glyph does *not* exist in the cache, we take the normal, "big"
font lock.

The lock is held during the entire glyph rasterization process, until
the new glyph has been inserted into the cache.

Only during the cache resize operation do we take a write-lock variant
of the rwlock, to ensure no other threads try to read the cache during
this time.