fcft: set LCD filter
FreeType has two different modes for subpixel antialiasing, selected
via a compile-time option: FT_CONFIG_OPTION_SUBPIXEL_RENDERING.

Despite its name, disabling this (the default, btw), does *not*
disable subpixel antialiasing, but causes FreeType to use its
own (newer), patent-free subpixel rendering mode named Harmony.

Enabling it causes FreeType to use the patent-encumbered ClearType
rendering mode.

The ClearType mode *requires* an LCD filter to be set, or the result
will have *severe* color fringes.

Now, since the API to set the filter is per *library* instance, we
need to re-set the filter before rendering a glyph, since different
fonts may have been configured to use different filters.

But, since it is per library instance, we need to guard both the call
to set the filter, *and* the call to render the glyph using the
library global lock.

This hurts performance, especially considering that most FreeType
installations does *not* use ClearType, and in this case the call to
set the LCD filter will actually fail.

To do the best we can, we detect at instantiation time whether we can
set a filter or not.

When rendering a glyph, we only set the filter if we can, *and* if we
need to (i.e we doing subpixel antialiased rendering).