fcft: fini: pop font from cache before calling fcft_destroy()
fcft_destroy() normally removes the font from the cache. But, if the
font failed to load, and the cache entry only contains a NULL pointer,
there is nothing for fcft_destroy() to remove.

Removing it directly in fini() solves this problem. fcft_destroy()
will still scan the cache looking for a matching font to remove. It
will fail, but this is not a problem since it still proceeds with
destroying the font.

Fixes https://codeberg.org/dnkl/foot/issues/174