fcft: text-shaping support
This adds initial, basic, support for text-shaping of whole text
runs.

It uses HarfBuzz to shape strings, using this algorithm:

* (Try to) shape the entire string, using the primary font.
* Parse the result, and at the first point of failure (i.e. first
  codepoint it failed to shape), re-start shaping using the first
  fallback font.
* If that still fails, try with the next fallback font
* As soon we’ve found a working fallback font, revert back to the
  primary font and continue (re-start) with the next codepoint.

RTL (right-to-left) scripts are semi-supported; we use HarfBuzz’s
auto-detection each time we’re trying to shape a segment. This means
the result of mixed RTL/LTR strings depend on the font being used; if
the font has glyphs for all the codepoints (and the entire string can
thus be shaped in a single go), all resulting glyphs will be either
RTL or LTR, depending on HarfBuzz’s initial guess.

If the font does *not* contain all glyphs, the RTL/LTR property will
be reavaluated each time we switch font. In this case, sub-sets of the
resulting glyphs may be RTL while others are LTR.