fcft: round, instead of truncating, the pixel size value passed to FreeType
FontConfig calculates a pixel size value from a point size + DPI. This
pixel size is a floating point value.

We used to pass it as-is to FreeType (when setting the font
size). This meant the value was truncated.

Now we round that value instead. So 11.8px now becomes 12px, instead
of 11px.