fcft: glyph/grapheme cache: fix rare crash
If one thread was resizing the glyph/grapheme cache, another thread
doing a cache lookup could, in rare cases, crash.

The cache lookup code does this:

    size_t idx = glyph_hash_index(font, hash_value_for_wc(wc, subpixel));
    struct glyph_priv **glyph = &font->glyph_cache.table[idx];

    while (*glyph != NULL && !((*glyph)->public.wc == wc &&
                               (*glyph)->subpixel == subpixel))
    {
        idx = (idx + 1) & (font->glyph_cache.size - 1);
        glyph = &font->glyph_cache.table[idx];
    }

I.e it calculates a hash index, and then loops until it finds a match,
or a NULL pointer entry (that is, the glyph doesn’t exist in the
cache).

Now, the collision handling is simple: just check that the properties
of the entry matches, and if not, check the next entry in the cache.

The cache resize function builds a new, larger hash, and copies over
the old cache entries to the new one, one by one. This is
done **without** the cache R/W lock, because we *want* other threads
to be able to do read-only cache lookups while we’re building the new
cache.

After creating the new hash table, the cache resize function does
this:

    pthread_rwlock_wrlock(&font->glyph_cache_lock);
    {
        free(font->glyph_cache.table);

        LOG_DBG("resized glyph cache from %zu to %zu", font->glyph_cache.size, size);
        font->glyph_cache.table = table;
        font->glyph_cache.size = size;
    }
    pthread_rwlock_unlock(&font->glyph_cache_lock);

I.e. it swaps out the old table with the new one, while holding the
R/W lock. This is all fine. The problem was in the loop that copies
the cache entries:

    for (size_t i = 0; i < font->glyph_cache.size; i++) {
        struct glyph_priv *entry = font->glyph_cache.table[i];

        if (entry == NULL)
            continue;

        size_t idx = hash_index_for_size(
            size, hash_value_for_wc(entry->public.wc, entry->subpixel));

        while (table[idx] != NULL) {
            assert(!(table[idx]->public.wc == entry->public.wc &&
                     table[idx]->subpixel == entry->subpixel));
            idx = (idx + 1) & (size - 1);
        }

        assert(table[idx] == NULL);
        table[idx] = entry;
        font->glyph_cache.table[i] = NULL;
    }

The last line *removes* the entry from the *current* cache. And
remember, this is done without holding any locks. This triggered a
NULL de-reference crash here:

    while (*glyph != NULL && !((*glyph)->public.wc == wc &&
                               (*glyph)->subpixel == subpixel))

where the *glyph != NULL check passed, but then the other thread
removed the entry from the table before the two remaining checks
ran. Oops.

The fix is simple: stop erasing the entries from the old cache. We
need to treat the old cache as read-only while not holding the R/W
lock.