fcft: destroy: don’t call fcft_destroy() on tll_front() if we just popped it
If loading a font fails, we still insert an entry in the font
cache. When destroying the font cache, we check if the cache entry is
NULL, in which case we just pop it from the cache.

And then we _also_ called fcft_destroy() on tll_front(). In many
cases, this actually works anyway. However, if we just popped
the *last* item, so that the cache is now empty, then tll_front() will
crash.