fcft: text-run rasterize: check codepoint is in font’s charset
A font may have it’s charset artificially limited (to prevent certain
glyphs from being used).

Since this is a FontConfig thing, it is up to us to check this before
attempting to load glyphs.

This is already being done by fcft_glyph_rasterize(), and
fcft_grapheme_rasterize(). But it was not being done in
fcft_text_run_rasterize().

text-run rasterization is different, since it attempts to shape
stretches of codepoints. Only at break points (i.e. codepoints which
the “current” font doesn’t have), do we get to apply any logic.

This patch deals with the charset map like fallback fonts in general
are handled; whenever we’re using a fallback font, the shaping logic
restarts with the primary font at the next grapheme cluster.

Closes #30