fcft: don’t bother with the charset if we’re forcing the primary font
If the shaping state is TEXT_RUN_PRIMARY_FORCE, then that means we’ve
already looped through *all* fonts (primary and fallbacks), and
couldn’t find a font with a glyph for the next codepoint.

In this case, we force the use of the primary font, even if HarfBuzz
says there’s no glyph available.

This means we should not skip the (primary) font, even if its charset
doesn’t include the next codepoint.