fcft: break the text-run if current font’s charset doesn’t include a codepoint
We already use FcCharSetHasChar() to avoid loading a fallback font if
said fallback font’s charset does not include the next codepoint to
shape.

However, if we’re shaping a substring that *contains* (but does
not *start* with) a codepoint that isn’t in the current font’s
charset, that went undetected and we still used the glyph from that
font (remember - a font may *have* a glyph, *even* if its charset says
it doesn’t).

This patch addresses this by, for each loaded glyph, check if the
corresponding cluster (or, first codepoint in that cluster) is in the
font’s charset map. If not, break the run (and try the next fallback
font).