fcft: refactor: add font_for_grapeme()
This function implements the logic for finding a font for a
grapheme.

It will try all fonts until it finds a font that has glyphs for all
codepoints in the grapheme. It also takes emoji presentation into
account; both the default (i.e. the implicit presentation selector),
and explicit selectors (0xfe0e and 0xfe0f).

If no font is found, it returns the primary font anyway.