fcft: text-run-rasterize: handle grapheme breaks and script changes
This is a rewrite of the algorithm used to shape and rasterize whole
text runs.

Previously, we let harfbuzz “consume” as much as possible of the text
run before even trying to change font.

I.e. we’d look at the next codepoint, find a font that has it, and
then let harfbuzz shape as much of the string as possible, using that
font.

Whenever harfbuzz failed, we’d go back and try the next font. There
are several problems with this:

* We only do LTR vs. RTL checks at these break points. This means that
  if a font has glyphs for many different scripts, we could, for
  example, end up reversing a LTR sub-string in the middle of other RTL
  sub-strings.

* We can’t handle emoji presentation selectors

* We may not be using the most appropriate font for parts of the
  string. I.e. the user may have configured FontConfig to use one font
  for Arabic, and another one for Hebrew. If those fonts also have
  glyphs for the other script, we’d end up using only one of the
  fonts.

The new algorithm takes a completely different approach. Before trying
to shape the run, we do grapheme cluster segmentation using utf8proc.

Said graphemes are then mapped to a font in *exactly* the same way
fcft_grapheme_rasterize() does (they’re now using the same code).

After that, we merge the graphemes into partial text-runs. Two
consecutive runs can be merged if:

* They use the same font instance.
* They have the same script.

If neither is true, they are treated as separate runs, and shaped
individually.

By using the same code for mapping a grapheme to a font instance as
fcft_grapheme_rasterize(), we also gain the ability to handle emoji
presentation selectors (including their default presentation) for
“free”.