fcft: grapheme_rasterize: try again without forcing emoji presentation
If we fail to find a font for an emoji, try again without forcing its
presentation style.

That is, if we want to render an emoji in its text presentation style,
but fail to find a non-emoji font, we re-search all fonts again, but
this time ignoring presentation style, meaning we can now match emoji
fonts too. And vice versa.