generate-emoji-data: compact the array by storing ranges
Once we’ve gathered all emoji data, we merge consecutive codepoints,
with identical properties, into a single entry, representing a range
of codepoints.

The first codepoint is stored as is, in the ‘cp’ member. A new member,
‘count’, has been added, where we store the *number* of consecutive
codepoints covered by this entry.

A count of 1 means we have exactly one codepoint. I.e. the range’s
start == end.

An alternative would have been to store the first and last codepoints
in the range. However, that would require two 4-byte members, whereas
the solution chosen here requires 4+1 bytes.

The emoji_compare() function, used when bsearch:ing the emoji data,
has been re-written to account for this.

Furthermore, it has been refactored so that the key is now the
codepoint only, and not a pointer to a ‘struct emoji’ entry.

This is perfectly ok. I.e. the ‘a’ and ‘b’ arguments in the bsearch
compare function don’t have to be of the same type. The compare
function is *always* called with the key in argument ‘a’, and an array
entry (pointer) in argument ‘b’.

This brings down the size of emoji data from ~5.7K to ~1.3K.