svg: use image bounds for size, not the indicated size
Not sure if this is a nanosvg bug or not... but it appears that if any
of the image’s shapes have negative bounds, the size indicated in
NSVGimage is wrong.

Workaround by finding the min/max bounds ourselves, and calculating
the size from that.

This fixes an issue where Noto Color Emoji glyphs were oversized.

Next, apply a scaled offset when rasterizing the glyph. This offset is
also based on the shapes’ bounds. Basically, if the image has negative
bounds, we need to offset the rasterizer by an equal amount. For
positive (non-zero) bounds, we need to use a negative offset in the
rasterizer.

This fixes an issue where Noto Color Emoji glyphs were incorrectly
positioned (and thus clipped).