svg: make SVG hooks thread safe
The ‘state’ variable passed around the SVG hooks is *global*. That is,
FreeType calls ‘svg_init()’ *once*, when the *first* SVG glyph is
about to be rasterized.

Again, it is *global*. NOT per font.

This means we cannot use it, since we want to be able to rasterize
glyphs in parallel, as long as they belong to different fonts (we _do_
have locking in place to deal with that).

So, stop using the ‘state’ variable, and instead store our in-between
hooks state in the ‘generic’ field in the glyph slot. This field
is *not* used by FreeType, but intended to be used by client
applications.